+++ positronic/cfg/policy/action.py
@@ -4,17 +4,17 @@ from positronic import geom
 
 @cfn.config(rotation_representation=geom.Rotation.Representation.ROTVEC, offset=1)
 def relative_robot_position(rotation_representation: geom.Rotation.Representation, offset: int):
-    from positronic.inference.action import RelativeRobotPositionAction
+    from positronic.policy.action import RelativeRobotPositionAction
     return RelativeRobotPositionAction(offset=offset, rotation_representation=rotation_representation)
 
 
 @cfn.config(rotation_representation=geom.Rotation.Representation.QUAT)
 def absolute_position(rotation_representation: geom.Rotation.Representation):
-    from positronic.inference.action import AbsolutePositionAction
+    from positronic.policy.action import AbsolutePositionAction
     return AbsolutePositionAction(rotation_representation=rotation_representation)
 
 
 @cfn.config()
 def absolute_joint_position():
-    from positronic.inference.action import AbsoluteJointPositionAction
+    from positronic.policy.action import AbsoluteJointPositionAction
     return AbsoluteJointPositionAction()
