+++ utilities/sync_policy_runner.py
@@ -7,12 +7,12 @@ import torch
 from tqdm import tqdm
 
 import configuronic as cfn
-import positronic.cfg.inference.action
-import positronic.cfg.inference.policy
-import positronic.cfg.inference.state
+import positronic.cfg.policy.action
+import positronic.cfg.policy.policy
+import positronic.cfg.policy.observation
 import positronic.cfg.simulator
-from positronic.inference.action import ActionDecoder
-from positronic.inference.state import ObservationEncoder
+from positronic.policy.action import ActionDecoder
+from positronic.policy.observation import ObservationEncoder
 from positronic.run_inference import rerun_log_action, rerun_log_observation
 from positronic.simulator.mujoco.sim import MujocoSimulatorEnv
 
@@ -112,9 +112,9 @@ def run_policy_in_simulator(  # noqa: C901  Function is too complex
 run = cfn.Config(
     run_policy_in_simulator,
     env=positronic.cfg.simulator.simulator,
-    state_encoder=positronic.cfg.inference.state.end_effector_back_front,
-    action_decoder=positronic.cfg.inference.action.relative_robot_position,
-    policy=positronic.cfg.inference.policy.act,
+    state_encoder=positronic.cfg.policy.observation.end_effector_back_front,
+    action_decoder=positronic.cfg.policy.action.relative_robot_position,
+    policy=positronic.cfg.policy.policy.act,
     rerun_path="rerun.rrd",
     inference_time_sec=10,
     observation_hz=60,
